vue back server modify

+++ mybatisproj/src/main/java/com/robin/basis/controller/system/TenantInfoController.java
@@ -23,7 +23,7 @@ public class TenantInfoController extends AbstractMyBatisController<ITenantInfoS
     @PostMapping
     @PreAuthorize("@checker.isSuperAdmin()")
     public Map<String,Object> save(@RequestPart("tenant") TenantInfoDTO dto,@RequestPart(required = false) MultipartFile logo){
-        int count=service.lambdaQuery().eq(TenantInfo::getTenantCode,dto.getTenantCode()).eq(TenantInfo::getStatus, Const.VALID).count();
+        long count=service.lambdaQuery().eq(TenantInfo::getTenantCode,dto.getTenantCode()).eq(TenantInfo::getStatus, Const.VALID).count();
         if(count>0){
             return wrapFailedMsg("存在编码相同的企业!");
         }
